Changed the border for the image map workspace to be transparent so that mouse events work correctly again. changed chem analysis distance in image map to be saved in mm. Changed database to store doubles instead of ints for the referenceX,Y values. Run the 12_2_09.sql migration. fixed a couple bugs with adding/removing chem analyses points to the image map. zooming in and out correctly updates the chem analysis points position (some rounding error needs to be removed, though). chem analysis points are correctly displayed when the image map is first loaded. Previously they would show up only when you added them.
changed GeoReference count to only count GeoReferences with samples associated to them.
fixed the count of private entities displayed on the home page (security filters really do work).

+++ mpdb-common/src/edu/rpi/metpetdb/client/model/ImageOnGrid.java
@@ -5,8 +5,8 @@ public class ImageOnGrid extends MObject {
 	private int id;
 	private Grid grid;
 	private Image image;
-	private int topLeftX;
-	private int topLeftY;
+	private double topLeftX;
+	private double topLeftY;
 	private int zOrder;
 	private int opacity;
 	private double resizeRatio;
@@ -80,20 +80,20 @@ public class ImageOnGrid extends MObject {
 		gchecksumHalf = s;
 	}
 
-	public int getTopLeftX() {
+	public double getTopLeftX() {
 		return topLeftX;
 	}
 
-	public void setTopLeftX(final int s) {
+	public void setTopLeftX(final double s) {
 		topLeftX = s;
 		// originalTopLeftX = s;
 	}
 
-	public int getTopLeftY() {
+	public double getTopLeftY() {
 		return topLeftY;
 	}
 
-	public void setTopLeftY(final int s) {
+	public void setTopLeftY(final double s) {
 		topLeftY = s;
 		// originalTopLeftY = s;
 	}
